requirements.txt
discord.py
datetime
google-generativeai
pymongo
python-dotenv
opencv-python
yt-dlp
urllib